Fix warning ⚠

+++ kratos/solving_strategies/builder_and_solvers/residualbased_block_builder_and_solver.h
@@ -1583,7 +1583,7 @@ protected:
     {
         double diagonal_norm = 0.0;
         #pragma omp parallel for reduction(+:diagonal_norm)
-        for(int i = 0; i < TSparseSpace::Size1(rA); ++i) {
+        for(int i = 0; i < static_cast<int>(TSparseSpace::Size1(rA)); ++i) {
             diagonal_norm += std::pow(rA(i,i), 2);
         }
         return std::sqrt(diagonal_norm);
@@ -1609,7 +1609,7 @@ protected:
 //         // NOTE: Reduction failing in MSVC
 //         double max_diag = 0.0;
 //         #pragma omp parallel for reduction(max:max_diag)
-//         for(IndexType i = 0; i < TSparseSpace::Size1(rA); ++i) {
+//         for(int i = 0; i < static_cast<int>(TSparseSpace::Size1(rA)); ++i) {
 //             max_diag = std::max(max_diag, std::abs(rA(i,i)));
 //         }
 //         return max_diag;
@@ -1618,7 +1618,7 @@ protected:
         const int num_threads = OpenMPUtils::GetNumThreads();
         Vector max_vector(num_threads, 0.0);
         #pragma omp parallel for
-        for(int i = 0; i < TSparseSpace::Size1(rA); ++i) {
+        for(int i = 0; i < static_cast<int>(TSparseSpace::Size1(rA)); ++i) {
             const int id = OpenMPUtils::ThisThread();
             if (std::abs(rA(i,i)) > max_vector[id])
                 max_vector[id] = std::abs(rA(i,i));
@@ -1641,7 +1641,7 @@ protected:
 //         // NOTE: Reduction failing in MSVC
 //         double min_diag = std::numeric_limits<double>::max();
 //         #pragma omp parallel for reduction(min:min_diag)
-//         for(IndexType i = 0; i < TSparseSpace::Size1(rA); ++i) {
+//         for(int i = 0; i < static_cast<int>(TSparseSpace::Size1(rA)); ++i) {
 //             min_diag = std::min(min_diag, std::abs(rA(i,i)));
 //         }
 //         return min_diag;
@@ -1650,7 +1650,7 @@ protected:
         const int num_threads = OpenMPUtils::GetNumThreads();
         Vector min_vector(num_threads, std::numeric_limits<double>::max());
         #pragma omp parallel for
-        for(int i = 0; i < TSparseSpace::Size1(rA); ++i) {
+        for(int i = 0; i < static_cast<int>(TSparseSpace::Size1(rA)); ++i) {
             const int id = OpenMPUtils::ThisThread();
             if (std::abs(rA(i,i)) < min_vector[id])
                 min_vector[id] = std::abs(rA(i,i));
